Sync, Sound & Mappers
- Implemented sound envelope for channel 1 & 2
- Fixed length timer being still incremented after max value causing strange behavior
- The Game reading is now handled by a specialized function in CartridgeHelper.h
- The CPU and co are now synced to the display framerate (WIP)
- Fixed Timer counting incorrectly
- uint is undefined on Windows
- Made some changed on the compilation steps

+++ hardware/Timer.h
@@ -44,7 +44,7 @@ private:
     class Bus *mBus;
 
     uint16_t internalCounter;
-    inline static std::array<uint, 4> clockSpeed = {256, 4, 16, 64};
+    inline static std::array<uint16_t, 4> clockSpeed = {256, 4, 16, 64};
 };
